Move config writer into Parse
