+++ src/Parse/Contracts/FileInterface.php
