+++ src/Parse/EnvFile.php
