+++ src/Parse/PHP/PHPConstant.php
